Refactor tests for components and messages
