+++ fe1-web/model/network/method/message/data/chirp/__tests__/NotifyAddChirp.test.ts
@@ -3,8 +3,8 @@ import '__tests__/utils/matchers';
 
 import { Base64UrlData, Hash, Timestamp } from 'model/objects';
 import { ProtocolError } from 'model/network/ProtocolError';
-import { NotifyAddChirp } from '../chirp';
-import { ActionType, ObjectType } from '../MessageData';
+import { NotifyAddChirp } from '../index';
+import { ActionType, ObjectType } from '../../MessageData';
 
 const TIMESTAMP = new Timestamp(1609455600); // 1st january 2021
 const CHANNEL = '/root/laoID/social/senderPublicKey';
