+++ fe1-web/model/network/method/message/data/reaction/__tests__/AddReaction.test.ts
@@ -3,8 +3,8 @@ import '__tests__/utils/matchers';
 
 import { Base64UrlData, Hash, Timestamp } from 'model/objects';
 import { ProtocolError } from 'model/network/ProtocolError';
-import { AddReaction } from '../reaction';
-import { ActionType, ObjectType } from '../MessageData';
+import { AddReaction } from '../index';
+import { ActionType, ObjectType } from '../../MessageData';
 
 const TIMESTAMP = new Timestamp(1609455600); // 1st january 2021
 const mockMessageId = Base64UrlData.encode('message_id');
